% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-tagging.R
\name{get_todays_sample_tweets}
\alias{get_todays_sample_tweets}
\title{Get a sample of latest tweet geolocations (deprecated)}
\usage{
get_todays_sample_tweets(limit = 1000, text_col = "text", lang_col = "lang")
}
\arguments{
\item{limit}{Size of the sample, default: 100}

\item{text_col}{Name of the tweet field to geolocate it should be one of the following ("text", "linked_text", "user_description", "user_location", "place_full_name", "linked_place_full_name"),
default: 'text'}

\item{lang_col}{Name of the tweet variable containing the language to evaluate. It should be one of the following ("lang", "linked_lang", NA), default: "lang"}
}
\value{
Data frame containing the sampled tweets and the geolocation metrics
}
\description{
This function was removed from epitweetr v1.0.1. Please use search_tweets instead
}
\details{
This function was removed from epitweetr v1.0.1. Please use search_tweets instead.
}
\examples{
if(FALSE){
   library(epitweetr)
   # setting up the data folder
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())

   # geolocating today's tweets
   show(get_todays_sample_tweets())
}
}
\seealso{
\code{\link{download_dependencies}}

 \code{\link{update_geonames}}

 \code{\link{update_languages}}
}
