% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{health_check}
\alias{health_check}
\title{Send email to administrator if a failure of epitweetr is detected}
\usage{
health_check(send_mail = TRUE, one_per_day = TRUE)
}
\arguments{
\item{send_mail}{Boolean. Whether an email should be sent to the defined administrator, default: TRUE}

\item{one_per_day}{Boolean. Whether a limit of one email per day will be applied, default: TRUE}
}
\value{
A list of health check errors found
}
\description{
It validates if epitweetr is not collecting tweets, aggregating tweets or not calculating alerts
}
\details{
This function sends an email to the defined administrator if epitweetr is not collecting tweets, aggregating tweets or not calculating alerts
}
\examples{
if(FALSE){
   #importing epitweetr
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   #sending the email to the administrator if epitweetr components are not properly working
   health_check()
}
}
