% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{is_detect_running}
\alias{is_detect_running}
\title{Check whether the alert detection task is running}
\usage{
is_detect_running()
}
\value{
logical Whether the alert detection task is running
}
\description{
gets the alert detection runner execution status
}
\details{
returns a logical value being TRUE if the alert detection task is actually running
}
\examples{
if(FALSE){
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   is_detect_running()
}
}
