% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{is_fs_running}
\alias{is_fs_running}
\title{Check whether the database is running}
\usage{
is_fs_running()
}
\value{
logical Whether the database is running
}
\description{
gets the database runner execution status
}
\details{
returns a logical value being TRUE if the database runner is actually running
}
\examples{
if(FALSE){
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   is_fs_running()
}
}
