% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{is_search_running}
\alias{is_search_running}
\title{Check whether the tweet collection task is running}
\usage{
is_search_running()
}
\value{
logical Whether the tweet collection is running
}
\description{
gets the tweet collection execution status
}
\details{
returns a logical value being TRUE if the tweet collection is running
}
\examples{
if(FALSE){
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   is_search_running()
}
}
