% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{register_detect_runner_task}
\alias{register_detect_runner_task}
\title{Registers the alert detection task}
\usage{
register_detect_runner_task()
}
\value{
Nothing
}
\description{
registers the alert detection task or stops if no configuration has been set or if it is already running
}
\details{
Registers the alert detection task or stops if no configuration has been set or if it is already running.
To generate alerts, this task needs the tweet collection to had successfully run since the last time it ran.
This function will use the task scheduler on windows and will fall back to launching the runner as a separate process (attached to this session) on Linux.
}
\examples{
if(FALSE){
   #getting tasks statuses
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   register_detect_runner_task()
}
}
