% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{register_fs_monitor}
\alias{register_fs_monitor}
\title{Registers the fs_monitor for the current process or exits}
\usage{
register_fs_monitor()
}
\value{
Nothing
}
\description{
registers the fs_monitor (by writing detect.PID file) for the current process or stops if no configuration has been set or if it is already running
}
\details{
Registers the fs_monitor (by writing detect.PID file) for the current process or stops if no configuration has been set or if it is already running
this function is exported so it can be called nicely from using the future package, but it is not intended to be directly used by users
}
\examples{
if(FALSE){
   #getting tasks statuses
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   register_fs_monitor()
}
}
