% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_twitter_app_auth}
\alias{set_twitter_app_auth}
\title{Save Twitter App credentials}
\usage{
set_twitter_app_auth(
  app = "",
  access_token = "",
  access_token_secret = "",
  api_key = "",
  api_secret = "",
  bearer = ""
)
}
\arguments{
\item{app}{Application name}

\item{access_token}{Access token as provided by Twitter}

\item{access_token_secret}{Access token secret as provided by Twitter}

\item{api_key}{API key as provided by Twitter}

\item{api_secret}{API secret as provided by Twitter}

\item{bearer}{the bearer token of the application}
}
\value{
Nothing
}
\description{
Update Twitter authentication tokens in a configuration object
}
\details{
Update Twitter authentication tokens in configuration object
}
\examples{
if(FALSE){
 #Setting the configuration values
   set_twitter_app_auth(
     app = "my super app", 
     access_token = "123456", 
     access_token_secret = "123456", 
     api_key = "123456", 
     api_secret = "123456"
   )
   set_twitter_app_auth(
     bearer = "123456"
   )
}
}
\seealso{
\code{\link{save_config}}
}
