% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{stop_search_runner_task}
\alias{stop_search_runner_task}
\title{Stops the tweet collection task}
\usage{
stop_search_runner_task()
}
\value{
Nothing
}
\description{
stops the tweet collection task
}
\details{
Stops the tweet collection task if it is already running
This function will try also deactivate the respective scheduled task on Windows.
}
\examples{
if(FALSE){
   #getting tasks statuses
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   stop_search_runner_task()
}
}
