% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_all}
\alias{check_all}
\title{Run automatic sanity checks}
\usage{
check_all()
}
\value{
Dataframe containing the statuses of all realized checks
}
\description{
It runs a set of automated sanity checks for helping the user to troubleqhot issues
}
\details{
This function executes a series of sanity checks, concerninr, java, bitness, task statusn dependencies and Twitter authentication.
}
\examples{
if(FALSE){
   #importing epitweer
   library(epitweetr)
   message('Please choose the epitweetr data directory')
   setup_config(file.choose())
   #running all tests
   check_all()
}
}
